backend/requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
python-dotenv
alembic